[NU-1979] FlinkDeploymentManager simpler - FlinkStatusDetailsDeterminer extracted
